requirements.txt
matplotlib==3.6.2
numpy==1.22.4
open3d==0.16.0
pynput==1.7.6
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1